[/] [SpawnNode] Will always set .Next to zero..
Before I was passing argument for .Next. But embedding element
in structure is not in function contract. It's job is allocate
list memory slot and set .Payload.

Strictly speaking setting .Payload is not in contract too.
But unlike .Next, there are no options for .Payload value.
So this boring code will propagate to Add's().